// src/structures/config.type.ts
type ConfigType = {
  id: String;
  appUUID: String;
  name: String;
  photoProfile: String;
  greeting: String;
  description: String;
  background: String;
  initialPrice: Number;
  finalPrice: Number;
  currency: String;
  address: String;
  webSite: String;
  urlKeybe: String;
  socialLinks: String;
  published: Boolean;
  createdAt: String;
  updateAt: String;
};

export default ConfigType;